requirements.txt
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.41
botocore==1.35.41
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
duckduckgo-search==6.3.2
email-validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
gitdb==4.0.11
gitpython==3.1.43
h11==0.14.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==3.0.1
mdurl==0.1.2
mistune==3.0.2
mypy==1.12.0
mypy-extensions==1.0.0
narwhals==1.9.3
nest-asyncio==1.6.0
numpy==2.1.2
openai==1.51.2
packaging==24.1
pandas==2.2.3
pgvector==0.3.5
phidata==2.5.3
pillow==10.4.0
pluggy==1.5.0
primp==0.6.4
protobuf==5.28.2
psycopg==3.1.19
psycopg-binary==3.1.19
pyarrow==17.0.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pydeck==0.9.1
pygments==2.18.0
pypdf==5.0.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pyyaml==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
ruff==0.6.9
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.36
starlette==0.40.0
streamlit==1.39.0
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
types-beautifulsoup4==4.12.0.20240907
types-html5lib==1.1.11.20240806
types-pillow==10.2.0.20240822
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
wikitextparser==0.56.3
